+++ packages/sdk/fastly/src/api/createCallbacks.ts
@@ -1,0 +1,9 @@
+const createCallbacks = () => ({
+  onError: () => {},
+  onFailed: () => {},
+  onReady: () => {},
+  onUpdate: () => {},
+  hasEventListeners: () => false,
+});
+
+export default createCallbacks;