+++ packages/shared/sdk-server/src/index.ts
@@ -9,6 +9,7 @@ export * from './store';
 export * from './events';
 
 export * from '@launchdarkly/js-sdk-common';
+export * as internalServer from './internal';
 
 export {
   LDClientImpl,