+++ packages/sdk/akamai-base/CHANGELOG.md
@@ -30,6 +30,53 @@ All notable changes to the LaunchDarkly SDK for Akamai Workers will be documente
     * @launchdarkly/akamai-edgeworker-sdk-common bumped from ^1.1.1 to ^1.1.2
     * @launchdarkly/js-server-sdk-common bumped from ^2.2.1 to ^2.2.2
 
+## [3.0.0](https://github.com/launchdarkly/js-core/compare/akamai-server-base-sdk-v2.1.23...akamai-server-base-sdk-v3.0.0) (2025-02-26)
+
+
+### ⚠ BREAKING CHANGES
+
+* Replace prefetch behavior with simple TTL cache ([#786](https://github.com/launchdarkly/js-core/issues/786))
+
+### Features
+
+* Replace prefetch behavior with simple TTL cache ([#786](https://github.com/launchdarkly/js-core/issues/786)) ([48b48cf](https://github.com/launchdarkly/js-core/commit/48b48cf69d518dc70a557ffd1dfb0209aee0b124))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/akamai-edgeworker-sdk-common bumped from ^1.4.1 to ^2.0.0
+
+## [2.1.23](https://github.com/launchdarkly/js-core/compare/akamai-server-base-sdk-v2.1.22...akamai-server-base-sdk-v2.1.23) (2025-02-18)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/akamai-edgeworker-sdk-common bumped from ^1.4.0 to ^1.4.1
+    * @launchdarkly/js-server-sdk-common bumped from ^2.11.0 to ^2.11.1
+
+## [2.1.22](https://github.com/launchdarkly/js-core/compare/akamai-server-base-sdk-v2.1.21...akamai-server-base-sdk-v2.1.22) (2025-01-30)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/akamai-edgeworker-sdk-common bumped from ^1.3.3 to ^1.4.0
+
+## [2.1.21](https://github.com/launchdarkly/js-core/compare/akamai-server-base-sdk-v2.1.20...akamai-server-base-sdk-v2.1.21) (2025-01-22)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/akamai-edgeworker-sdk-common bumped from ^1.3.2 to ^1.3.3
+    * @launchdarkly/js-server-sdk-common bumped from ^2.10.0 to ^2.11.0
+
 ## [2.1.20](https://github.com/launchdarkly/js-core/compare/akamai-server-base-sdk-v2.1.19...akamai-server-base-sdk-v2.1.20) (2024-11-14)