+++ packages/sdk/akamai-edgekv/CHANGELOG.md
@@ -31,6 +31,49 @@ All notable changes to the LaunchDarkly SDK for Akamai Workers will be documente
     * @launchdarkly/akamai-edgeworker-sdk-common bumped from ^1.1.1 to ^1.1.2
     * @launchdarkly/js-server-sdk-common bumped from ^2.2.1 to ^2.2.2
 
+## [1.4.2](https://github.com/launchdarkly/js-core/compare/akamai-server-edgekv-sdk-v1.4.1...akamai-server-edgekv-sdk-v1.4.2) (2025-02-26)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/akamai-edgeworker-sdk-common bumped from ^1.4.1 to ^2.0.0
+
+## [1.4.1](https://github.com/launchdarkly/js-core/compare/akamai-server-edgekv-sdk-v1.4.0...akamai-server-edgekv-sdk-v1.4.1) (2025-02-18)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/akamai-edgeworker-sdk-common bumped from ^1.4.0 to ^1.4.1
+    * @launchdarkly/js-server-sdk-common bumped from ^2.11.0 to ^2.11.1
+
+## [1.4.0](https://github.com/launchdarkly/js-core/compare/akamai-server-edgekv-sdk-v1.3.1...akamai-server-edgekv-sdk-v1.4.0) (2025-01-30)
+
+
+### Features
+
+* Add cacheTtlMs option ([#760](https://github.com/launchdarkly/js-core/issues/760)) ([4f961dd](https://github.com/launchdarkly/js-core/commit/4f961dd16fd10f5bb55dd2116d26b218944bfeb2))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/akamai-edgeworker-sdk-common bumped from ^1.3.3 to ^1.4.0
+
+## [1.3.1](https://github.com/launchdarkly/js-core/compare/akamai-server-edgekv-sdk-v1.3.0...akamai-server-edgekv-sdk-v1.3.1) (2025-01-22)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/akamai-edgeworker-sdk-common bumped from ^1.3.2 to ^1.3.3
+    * @launchdarkly/js-server-sdk-common bumped from ^2.10.0 to ^2.11.0
+
 ## [1.3.0](https://github.com/launchdarkly/js-core/compare/akamai-server-edgekv-sdk-v1.2.1...akamai-server-edgekv-sdk-v1.3.0) (2024-11-14)