+++ packages/sdk/cloudflare/CHANGELOG.md
@@ -21,6 +21,24 @@ All notable changes to the LaunchDarkly SDK for Cloudflare Workers will be docum
   * devDependencies
     * @launchdarkly/js-server-sdk-common-edge bumped from 2.2.1 to 2.2.2
 
+## [2.6.5](https://github.com/launchdarkly/js-core/compare/cloudflare-server-sdk-v2.6.4...cloudflare-server-sdk-v2.6.5) (2025-02-18)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/js-server-sdk-common-edge bumped from 2.5.3 to 2.5.4
+
+## [2.6.4](https://github.com/launchdarkly/js-core/compare/cloudflare-server-sdk-v2.6.3...cloudflare-server-sdk-v2.6.4) (2025-01-22)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/js-server-sdk-common-edge bumped from 2.5.2 to 2.5.3
+
 ## [2.6.3](https://github.com/launchdarkly/js-core/compare/cloudflare-server-sdk-v2.6.2...cloudflare-server-sdk-v2.6.3) (2024-12-12)