+++ packages/sdk/server-node/CHANGELOG.md
@@ -2,6 +2,29 @@
 
 All notable changes to `@launchdarkly/node-server-sdk` will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org).
 
+## [9.7.4](https://github.com/launchdarkly/js-core/compare/node-server-sdk-v9.7.3...node-server-sdk-v9.7.4) (2025-02-18)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/js-server-sdk-common bumped from 2.11.0 to 2.11.1
+
+## [9.7.3](https://github.com/launchdarkly/js-core/compare/node-server-sdk-v9.7.2...node-server-sdk-v9.7.3) (2025-01-22)
+
+
+### Bug Fixes
+
+* Fix typo in proxy-authorization header for basic authentication. ([#720](https://github.com/launchdarkly/js-core/issues/720)) ([220b6d6](https://github.com/launchdarkly/js-core/commit/220b6d6d34331d271ca30f0cae363c734fcc38bf)), closes [#718](https://github.com/launchdarkly/js-core/issues/718)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/js-server-sdk-common bumped from 2.10.0 to 2.11.0
+
 ## [9.7.2](https://github.com/launchdarkly/js-core/compare/node-server-sdk-v9.7.1...node-server-sdk-v9.7.2) (2024-11-14)