+++ packages/sdk/vercel/CHANGELOG.md
@@ -20,6 +20,24 @@ All notable changes to the LaunchDarkly SDK for Vercel Edge Config will be docum
   * dependencies
     * @launchdarkly/js-server-sdk-common-edge bumped from 2.2.1 to 2.2.2
 
+## [1.3.23](https://github.com/launchdarkly/js-core/compare/vercel-server-sdk-v1.3.22...vercel-server-sdk-v1.3.23) (2025-02-18)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/js-server-sdk-common-edge bumped from 2.5.3 to 2.5.4
+
+## [1.3.22](https://github.com/launchdarkly/js-core/compare/vercel-server-sdk-v1.3.21...vercel-server-sdk-v1.3.22) (2025-01-22)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/js-server-sdk-common-edge bumped from 2.5.2 to 2.5.3
+
 ## [1.3.21](https://github.com/launchdarkly/js-core/compare/vercel-server-sdk-v1.3.20...vercel-server-sdk-v1.3.21) (2024-11-14)