+++ packages/shared/sdk-server/CHANGELOG.md
@@ -8,6 +8,27 @@ All notable changes to `@launchdarkly/js-server-sdk-common` will be documented i
   * dependencies
     * @launchdarkly/js-sdk-common bumped from 2.3.0 to 2.3.1
 
+## [2.11.1](https://github.com/launchdarkly/js-core/compare/js-server-sdk-common-v2.11.0...js-server-sdk-common-v2.11.1) (2025-02-18)
+
+
+### Bug Fixes
+
+* Fix issue where flush callback could be called twice. ([#779](https://github.com/launchdarkly/js-core/issues/779)) ([c377e89](https://github.com/launchdarkly/js-core/commit/c377e890f9af71f1658f3303217118206496a602))
+
+## [2.11.0](https://github.com/launchdarkly/js-core/compare/js-server-sdk-common-v2.10.0...js-server-sdk-common-v2.11.0) (2025-01-22)
+
+
+### Features
+
+* Adds StreamingProcessor for FDv2 to sdk-server package. ([#707](https://github.com/launchdarkly/js-core/issues/707)) ([7f5c275](https://github.com/launchdarkly/js-core/commit/7f5c2750dcc8341d049d7e736ca21ec36e168703))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/js-sdk-common bumped from 2.12.0 to 2.13.0
+
 ## [2.10.0](https://github.com/launchdarkly/js-core/compare/js-server-sdk-common-v2.9.1...js-server-sdk-common-v2.10.0) (2024-11-14)