+++ packages/tooling/jest/CHANGELOG.md
@@ -1,3 +1,21 @@
 # Changelog
 
+## [0.1.1](https://github.com/launchdarkly/js-core/compare/jest-v0.1.0...jest-v0.1.1) (2025-02-06)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/react-native-client-sdk bumped from ~10.9.0 to ~10.9.6
+
+## 0.1.0 (2025-01-24)
+
+
+### Features
+
+* Implement jest mocks for react-native. ([#535](https://github.com/launchdarkly/js-core/issues/535)) ([aee09c8](https://github.com/launchdarkly/js-core/commit/aee09c87e1810795108157dcd9ab8cfb9fe05020))
+
+## Changelog
+
 All notable changes to the LaunchDarkly Jest package will be documented in this file. This project adheres to [Semantic Versioning](https://semver.org).